+++ CHANGELOG.md
@@ -127,4 +127,5 @@ Thanks to [UdHo](https://github.com/UdHo)
 * `Vector2::new_x()`
 * `Vector2::new_y()`
 * `Rectangle::with_negative_offset()`
+* `Rectangle::ZERO`
 * `Color::from_gray()`